created tile sets and obstacles
Created 2 diffrent tile sets, one with 3 tiles, making it kid of a platform and another one with a single one for enemies or jumping. Some obstacles such as spikes and saws have also been created to increase difficulty, however I was unable to add collisions to them. I followed the tutorial however, since I made it as a 2d tile set, it might not have implemented collision for the 3d game that it is now